% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AUC_base.R
\name{calculate_AUC_base}
\alias{calculate_AUC_base}
\title{Title Main}
\usage{
calculate_AUC_base(
  n01_p = 15,
  n_train_sets = c(15, 30, 60, 120, 150),
  n01_test = 300,
  seed = 1,
  method = "pca2_mvnorm",
  model = c("svm", "randomforest")
)
}
\arguments{
\item{n01_p}{size of pilot data labeled as class 0/1}

\item{n_train_sets}{size sets of training data labeled as class 0/1}

\item{n01_test}{size of test data labeled as class 0/1}

\item{seed}{random seed number}

\item{method}{Choose the method you want to use: "pca2_mvnorm" and "gaussian_copula".
The default value is "pca2_mvnorm".}

\item{model}{base classification model.
\itemize{
\item logistic: Logistic regression. \link{glm} function with family = 'binomial'
\item penlog: Penalized logistic regression with LASSO penalty. \code{\link[glmnet]{glmnet}} in \code{glmnet} package
\item svm: Support Vector Machines. \code{\link[e1071]{svm}} in \code{e1071} package
\item randomforest: Random Forest. \code{\link[randomForest]{randomForest}} in \code{randomForest} package
\item lda: Linear Discriminant Analysis. \code{\link[MASS]{lda}} in \code{MASS} package
\item slda: Sparse Linear Discriminant Analysis with LASSO penalty.
\item nb: Naive Bayes. \code{\link[e1071]{naiveBayes}} in \code{e1071} package
\item nnb: Nonparametric Naive Bayes. \code{\link[naivebayes]{naive_bayes}} in \code{naivebayes} package
\item ada: Ada-Boost. \code{\link[ada]{ada}} in \code{ada} package
\item xgboost: XGBboost. \code{\link[xgboost]{xgboost}} in \code{xgboost} package
\item tree: Classificatin Tree. \code{\link[tree]{tree}} in \code{tree} package
}}
}
\value{
AUC of ...
}
\description{
Title Main
}
\examples{
AUC = calculate_AUC_base(n01_p = 15, n01_test = 300, seed = 1)
}
