% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AUCs.R
\name{calculate_AUCs}
\alias{calculate_AUCs}
\title{Title Calculate the AUCs}
\usage{
calculate_AUCs(
  n01_p = 15,
  n_train_sets = c(15, 30, 60, 120, 150),
  n01_test = 300,
  num_of_seeds = 20,
  random_seeds = TRUE,
  seeds = NULL,
  calculate_std_of_AUC_and_produce_plot = FALSE,
  method = "pca2_mvnorm",
  ncores = NULL,
  model = c("svm", "randomforest")
)
}
\arguments{
\item{n01_p}{number of pilot data labeled as class 0/1}

\item{n_train_sets}{size sets of training data labeled as class 0/1}

\item{n01_test}{number of test data labeled as class 0/1}

\item{num_of_seeds}{number of seeds}

\item{random_seeds}{If TRUE, it will produce "num_of_seeds" seeds randomly.
If FASLE, it will use seeds "1:num_of_seeds", which is used to make it reproducible.}

\item{seeds}{You can choose the seeds set as you wish. It's default value is NULL.
If it's not NULL, the function will use your seeds and ignore parameters "num_of_seeds" and "random_seeds".}

\item{calculate_std_of_AUC_and_produce_plot}{You can calculate the standard deviation of AUC and produce the plot directly if you set this parameter to TRUE.}

\item{method}{Choose the method you want to use: "pca2_mvnorm" and "gaussian_copula".
The default value is "pca2_mvnorm".}

\item{ncores}{number of cores used for parallel computing. Default = detectCores() - 1.}

\item{model}{base classification model.
\itemize{
\item logistic: Logistic regression. \link{glm} function with family = 'binomial'
\item penlog: Penalized logistic regression with LASSO penalty. \code{\link[glmnet]{glmnet}} in \code{glmnet} package
\item svm: Support Vector Machines. \code{\link[e1071]{svm}} in \code{e1071} package
\item randomforest: Random Forest. \code{\link[randomForest]{randomForest}} in \code{randomForest} package
\item lda: Linear Discriminant Analysis. \code{\link[MASS]{lda}} in \code{MASS} package
\item slda: Sparse Linear Discriminant Analysis with LASSO penalty.
\item nb: Naive Bayes. \code{\link[e1071]{naiveBayes}} in \code{e1071} package
\item nnb: Nonparametric Naive Bayes. \code{\link[naivebayes]{naive_bayes}} in \code{naivebayes} package
\item ada: Ada-Boost. \code{\link[ada]{ada}} in \code{ada} package
\item xgboost: XGBboost. \code{\link[xgboost]{xgboost}} in \code{xgboost} package
\item tree: Classificatin Tree. \code{\link[tree]{tree}} in \code{tree} package
}}
}
\value{
Return the AUCs you want to calculate
}
\description{
Title Calculate the AUCs
}
\examples{
calculate_AUCs(n01_p = 15, n01_test = 300, num_of_seeds = 20, random_seeds = TRUE)
calculate_AUCs(n01_p = 15, n01_test = 300, num_of_seeds = 20, random_seeds = FALSE)
calculate_AUCs(n01_p = 15, n01_test = 300, seeds = c(2, 4, 6, 7, 9, 12, 25, 34, 24, 
    65))
calculate_AUCs(n01_p = 15, n01_test = 300, num_of_seeds = 20, random_seeds = TRUE, 
    calculate_std_of_AUC_and_produce_plot = TRUE)

}
