% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_std_of_AUC_and_draw_plot.R
\name{calculate_std_of_AUC_and_draw_plot}
\alias{calculate_std_of_AUC_and_draw_plot}
\title{Calculate std and draw the plots}
\usage{
calculate_std_of_AUC_and_draw_plot(res, n_train_sets, model)
}
\arguments{
\item{res}{the result produced by cal_parallel}

\item{n_train_sets}{size sets of training data labeled as class 0/1. For example, n_train_sets=c(c(30,30),c(90,90),c(150,150)) represents that we try 3 different sets of training data and the training size of the first set is c(30,30).}

\item{model}{base classification model set which is corresponding to the result in the "res".
\itemize{
\item logistic: Logistic regression. \link{glm} function with family = 'binomial'
\item penlog: Penalized logistic regression with LASSO penalty. \code{\link[glmnet]{glmnet}} in \code{glmnet} package
\item svm: Support Vector Machines. \code{\link[e1071]{svm}} in \code{e1071} package
\item randomforest: Random Forest. \code{\link[randomForest]{randomForest}} in \code{randomForest} package
\item lda: Linear Discriminant Analysis. \code{\link[MASS]{lda}} in \code{MASS} package
\item slda: Sparse Linear Discriminant Analysis with LASSO penalty.
\item nb: Naive Bayes. \code{\link[e1071]{naiveBayes}} in \code{e1071} package
\item nnb: Nonparametric Naive Bayes. \code{\link[naivebayes]{naive_bayes}} in \code{naivebayes} package
\item ada: Ada-Boost. \code{\link[ada]{ada}} in \code{ada} package
\item xgboost: XGBboost. \code{\link[xgboost]{xgboost}} in \code{xgboost} package
\item tree: Classificatin Tree. \code{\link[tree]{tree}} in \code{tree} package
}}
}
\value{
It will save all the result to the new "result" folder created in your current directory.
}
\description{
Calculate std and draw the plots
}
\examples{
AUC_1 = calculate_AUC_base(n01_p = c(15, 15), n01_test = c(300, 300), seed = 1, 
    model = c("svm", "randomforest"), n_train_sets = c(c(15, 15), c(30, 30), c(60, 
        60), c(120, 120), c(150, 150)))
AUC_2 = calculate_AUC_base(n01_p = c(15, 15), n01_test = c(300, 300), seed = 1, 
    model = c("svm", "randomforest"), n_train_sets = c(c(15, 15), c(30, 30), c(60, 
        60), c(120, 120), c(150, 150)))
res[[1]] = AUC_1
res[[2]] = AUC_2
calculate_std_of_AUC_and_draw_plot(res, n_train_sets = c(c(15, 15), c(30, 30), c(60, 
    60), c(120, 120), c(150, 150)), model = c("svm", "randomforest"))
}
