% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_data.R
\name{downsample_data}
\alias{downsample_data}
\title{Downsample Data}
\usage{
downsample_data(data_list, n0_train, n1_train, seed = NULL)
}
\arguments{
\item{data_list}{data list you want to downsample(including train_x, train_y and test_x)}

\item{n0_train}{Int, represents the absolute number of class 0 training data.}

\item{n1_train}{Int, represents the absolute number of class 1 training data.}
}
\value{
the downsampled data
}
\description{
Downsample the training data.
}
\examples{

library(mvtnorm)
library(MASS)

df = 10
rho = 0.5
d = 5
delta = rep(2, d)
H <- abs(outer(1:d, 1:d, "-"))
covxx = rho^H

n1_all <- n0_all <- 800
n1_p <- n0_p <- 15

x0_all = rmvt(n = n0_all, sigma = covxx, delta = rep(0, d), df = df)
x1_all = rmvt(n = n1_all, sigma = covxx, delta = delta, df = df)

x_data = rbind(x0_all, x1_all)
y_data = c(rep(0, n0_all), rep(1, n1_all))

id0 <- which(y_data == 0)
id1 <- which(y_data == 1)

id0_p <- sample(id0, n0_p)
id1_p <- sample(id1, n1_p)
id_p <- c(id0_p, id1_p)

x_pilot <- as.matrix(x_data[id_p, ])
y_pilot <- as.matrix(y_data[id_p])

n1_train <- n0_train <- n_train <- 150
n0_test <- n1_test <- 300

data_list = pilot_tfe_mvnorm_pca2(x_pilot, y_pilot, n0_train, n1_train, n0_test, 
    n1_test)

data_list_ds = downsample_data(data_list, n0_train = 30, n1_train = 30, seed = 1)

}
