% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{foo}
\alias{foo}
\title{Poisson vector.}
\usage{
foo(x1 = 2, x2 = 21)
}
\arguments{
\item{x1}{explain what is x1}

\item{x2}{explain what is x2}
}
\value{
explain what is returned by this function.
}
\description{
Creating \strong{poisson} vector.
}
\details{
Input an integer and return the log density of a poisson distribution with lambda equals the input integer.
}
\examples{
foo()
}
