% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_tfe_mvnorm_pca2.R
\name{pilot_tfe_mvnorm_pca2}
\alias{pilot_tfe_mvnorm_pca2}
\title{Estimation from Pilot Data using PCA_2 and Mvnorm}
\usage{
pilot_tfe_mvnorm_pca2(x_pilot, y_pilot, n0_train, n1_train, n0_test, n1_test)
}
\arguments{
\item{x_pilot}{input variables of pilot data}

\item{y_pilot}{labels of pilot data}

\item{n0_train}{the number of training data of class 0}

\item{n1_train}{the number of training data of class 1}

\item{n0_test}{the number of test data of class 0}

\item{n1_test}{the number of test data of class 1}
}
\value{
the scores predicted by Logistic Regresion, Random Forrest, Support Vector Machine and Xgboost seperately
}
\description{
Use PCA and select the first two PCs to reduce the dimension, then use multivariate normal distribution as the generative model and train 4 different models.
}
\details{
Use Logistic Regresion, Random Forrest, Support Vector Machine and Xgboost as training model.
}
\examples{

library(mvtnorm)
library(MASS)

df = 10
rho = 0.5
d = 5
delta = rep(2, d)
H <- abs(outer(1:d, 1:d, "-"))
covxx = rho^H

n1_all <- n0_all <- 800
n1_p <- n0_p <- 15

x0_all = rmvt(n = n0_all, sigma = covxx, delta = rep(0, d), df = df)
x1_all = rmvt(n = n1_all, sigma = covxx, delta = delta, df = df)

x_data = rbind(x0_all, x1_all)
y_data = c(rep(0, n0_all), rep(1, n1_all))

id0 <- which(y_data == 0)
id1 <- which(y_data == 1)

id0_p <- sample(id0, n0_p)
id1_p <- sample(id1, n1_p)
id_p <- c(id0_p, id1_p)

x_pilot <- as.matrix(x_data[id_p, ])
y_pilot <- as.matrix(y_data[id_p])

n1_train <- n0_train <- n_train <- 60
n0_test <- n1_test <- 300

result = pilot_tfe_mvnorm_pca2(x_pilot, y_pilot, n0_train, n1_train, n0_test, n1_test)



}
