% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Summarizes data}
\usage{
summarySE(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{data}{a data frame}

\item{measurevar}{the name of a column that contains the variable to be summariezed}

\item{groupvars}{a vector containing names of columns that contain grouping variables}

\item{na.rm}{a boolean that indicates whether to ignore NA's}

\item{conf.interval}{the percent range of the confidence interval (default is 95\%)}

\item{.drop}{should combinations of variables that do not appear in the input data be preserved (FALSE) or dropped (TRUE, default)}
}
\value{
the count, mean, standard deviation, standard error of the mean, and confidence interval of data
}
\description{
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).
}
\examples{

mean_values = c(0, 5)
sd_values = c(1, 5)

data <- data.frame(data.norm = c(rnorm(n = 50, mean = 0, sd = 1), rnorm(n = 50, 
    mean = 5, sd = 1), rnorm(n = 50, mean = 0, sd = 5), rnorm(n = 50, 
    mean = 5, sd = 5)), Mean_value = as.factor(rep(mean_values, each = 50)), 
    Sd_value = as.factor(rep(sd_values, each = 100)))

data.norm.sum <- summarySE(data, measurevar = "data.norm", groupvars = c("Mean_value", 
    "Sd_value"))
}
