% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AUC_base.R
\name{calculate_AUC_base}
\alias{calculate_AUC_base}
\title{Title Main}
\usage{
calculate_AUC_base(
  n01_p = 15,
  n_train_sets = c(15, 30, 60, 120, 150),
  n01_test = 300,
  seed = 1,
  method = "pca2_mvnorm"
)
}
\arguments{
\item{n01_p}{size of pilot data labeled as class 0/1}

\item{n_train_sets}{size sets of training data labeled as class 0/1}

\item{n01_test}{size of test data labeled as class 0/1}

\item{seed}{random seed number}

\item{method}{Choose the method you want to use: "pca2_mvnorm" and "gaussian_copula".
The default value is "pca2_mvnorm".}
}
\value{
AUC of ...
}
\description{
Title Main
}
\examples{
AUC = calculate_AUC(n01_p = 15, n01_test = 300, seed = 1)
}
