% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_std_of_AUC_and_draw_plot.R
\name{calculate_std_of_AUC_and_draw_plot}
\alias{calculate_std_of_AUC_and_draw_plot}
\title{Calculate std and draw the plots}
\usage{
calculate_std_of_AUC_and_draw_plot(res, n_train_sets = c(15, 30, 60, 120, 150))
}
\arguments{
\item{res}{the result produced by cal_parallel}
}
\value{
It will save all the result to the "result" directory.
}
\description{
Calculate std and draw the plots
}
\examples{
calculate_std_of_AUC_and_draw_plot(res)
}
