% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate Data}
\usage{
generate_data(seed = 1)
}
\arguments{
\item{seed}{the seed number}
}
\value{
the generated data
}
\description{
Generate Data
}
\examples{
generate_data(seed = 1)
}
