% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split Data}
\usage{
split_data(x_data, y_data, n_train = 0.7, n_test = NULL, seed = NULL)
}
\arguments{
\item{x_data}{x data you want to split}

\item{y_data}{y data you want to split: it must be binary data and the class label should be "0" and "1"}

\item{n_train}{If float, should be between 0.0 and 1.0 and represent the proportion of the dataset to include in the train split.
If int, represents the absolute number of train samples (It will extract half from class 0 and half from class 1).
If None, the value is automatically set to the complement of the test size.
Its default value is set to 0.7}

\item{n_test}{If float, should be between 0.0 and 1.0 and represent the proportion of the dataset to include in the train split.
If int, represents the absolute number of train samples (It will extract half from class 0 and half from class 1).
If NULL, the value is automatically set to the complement of the training size.}

\item{seed}{Controls the shuffling applied to the data before applying the split.
Pass an int as the seed for reproducible output across multiple function calls.}
}
\value{
the split data
}
\description{
Split Data
}
\examples{

data = generate_data()
data_split1 = split_data(data$x_data, data$y_data, n_train = 30, seed = 1)
data_split2 = split_data(data$x_data, data$y_data, n_train = 1000, n_test = 500, 
    seed = 1)
data_split3 = split_data(data$x_data, data$y_data, n_train = 0.6, n_test = 0.4, seed = 1)
}
